Adding Explicit Tests for Power State
Added more explicit tests to validate each power state. Also,
restructured PDU unit tests to inherit from base PDU test cases.
So each PDU type will share common functionality to not repeat
similar tests.
