+++ virtualpdu/tests/unit/pdu/test_base_pdu.py
@@ -14,19 +14,12 @@
 
 from mock import mock
 from virtualpdu.pdu import PDU
-from virtualpdu.power_states import POWER_OFF
 from virtualpdu.tests import base
+from virtualpdu.tests.unit.pdu.base_pdu_test_cases import BasePDUTests
 
 
-class TestPDU(base.TestCase):
+class TestPDU(base.TestCase, BasePDUTests):
     def setUp(self):
         super(TestPDU, self).setUp()
         self.core_mock = mock.Mock()
         self.pdu = PDU(name='my_pdu', core=self.core_mock)
-
-    def test_power_off_notifies_core(self):
-        self.pdu.oids[0].value = 'off'
-        self.core_mock.pdu_outlet_state_changed.assert_called_with(
-            name='my_pdu',
-            outlet_number=1,
-            state=POWER_OFF)
